+++ pybamm/solvers/scikits_dae_solver.py
@@ -54,7 +54,7 @@ class ScikitsDaeSolver(pybamm.BaseSolver):
         atol=1e-6,
         root_method="casadi",
         root_tol=1e-6,
-        extrap_tol=0,
+        extrap_tol=None,
         extra_options=None,
     ):
         if scikits_odes_spec is None: